Fixes #191: Skip Explorer restart when in sysprep mode
- Since Settings are applied to Default Profile, calling RestartExplorer
  does nothing
- RestartExplorer also terminates all processes witch might have called
  WinDebloat in the first place

+++ Win11Debloat.ps1
@@ -547,14 +547,14 @@ function RegImport {
 function RestartExplorer {
     Write-Output "> Restarting Windows Explorer process to apply all changes... (This may cause some flickering)"
 
-    # Only restart if the powershell process matches the OS architecture
+    # Only restart if the powershell process matches the OS architecture and not not in sysprep mode.
     # Restarting explorer from a 32bit Powershell window will fail on a 64bit OS
-    if ([Environment]::Is64BitProcess -eq [Environment]::Is64BitOperatingSystem)
+    if ( ([Environment]::Is64BitProcess -eq [Environment]::Is64BitOperatingSystem) -and (-not $global:Params.ContainsKey("Sysprep")) )
     {
         Stop-Process -processName: Explorer -Force
     }
     else {
-        Write-Warning "Unable to restart Windows Explorer process, please manually restart your PC to apply all changes."
+        Write-Warning "Not restarting Windows Explorer process, please manually restart your PC to apply all changes."
     }
 }
